rag-agent-pinecone-main/rag-agent-pinecone-main/requirements_streamlit.txt
streamlit>=1.28.0
pinecone-client>=2.2.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
